seguimos viendo, no funciona

+++ lucasColaboracion/TrabajoIntegradorTramo2/src/main/java/com/grupo3ex5/argprograma/trabajointegradortramo2/entidades/Tecnico.java
@@ -35,7 +35,7 @@ public class Tecnico implements Serializable {
         this.nombreApellido = nombreApellido;
     }
 
-    public long getId_tecnico() {
+    public int getId_tecnico() {
         return id_tecnico;
     }
